+++ src/main/java/com/example/TicketChain/repository/OrderDetailRepository.java
@@ -1,11 +1,13 @@
 package com.example.TicketChain.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.example.TicketChain.entity.OrderDetail;
 
 @Repository
-public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long> {
+public interface OrderDetailRepository extends JpaRepository<OrderDetail, BigInteger> {
 
 }
