+++ src/main/java/com/example/TicketChain/repository/TicketListingRepository.java
@@ -1,11 +1,13 @@
 package com.example.TicketChain.repository;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import com.example.TicketChain.entity.TicketListing;
 
 @Repository
-public interface TicketListingRepository extends JpaRepository<TicketListing, Long> {
+public interface TicketListingRepository extends JpaRepository<TicketListing, BigInteger> {
 
 }
